Add the i-slint-editor-preview crate with the document model
Move common.rs, common/ (document cache, text edits, EditorSession,
LspToPreviews with the RemoteTransport trait, the rename engines,
component catalog, token info, import edits, test fixtures) and
util.rs from tools/lsp into the new publishable crate
internal/editor-preview. The LSP re-exports the moved modules from its
crate roots, so existing paths keep resolving.

LSP-client code stays behind: the lsp_to_editor helpers and
publish_diagnostics move to tools/lsp/lsp_to_editor.rs and
host_language_search becomes a top-level LSP module, taking the
host-language rename tests with it.

The shared test fixtures sit behind a new testing feature, which the
LSP enables from its dev-dependencies. The preview feature ladder is
mirrored on the new crate and forwarded from slint-lsp.
