+++ Cargo.toml
@@ -38,6 +38,7 @@ members = [
   'internal/compiler/parser-test-macro',
   'internal/core',
   'internal/core-macros',
+  'internal/editor-preview',
   'internal/interpreter',
   'internal/live-preview',
   'tools/compiler',
@@ -92,6 +93,7 @@ i-slint-common = { version = "=1.18.0", path = "internal/common", default-featur
 i-slint-compiler = { version = "=1.18.0", path = "internal/compiler", default-features = false }
 i-slint-core = { version = "=1.18.0", path = "internal/core", default-features = false }
 i-slint-core-macros = { version = "=1.18.0", path = "internal/core-macros", default-features = false }
+i-slint-editor-preview = { version = "=1.18.0", path = "internal/editor-preview", default-features = false }
 i-slint-live-preview = { version = "=1.18.0", path = "internal/live-preview", default-features = false }
 i-slint-renderer-anyrender = { version = "=1.18.0", path = "internal/renderers/anyrender", default-features = false }
 i-slint-renderer-femtovg = { version = "=1.18.0", path = "internal/renderers/femtovg", default-features = false }
