+++ docs/development/lsp-architecture.md
@@ -31,9 +31,9 @@ The Slint LSP (Language Server Protocol) server provides IDE features for `.slin
 | `tools/lsp/language/hover.rs` | Hover information |
 | `tools/lsp/language/semantic_tokens.rs` | Syntax highlighting |
 | `tools/lsp/language/signature_help.rs` | Function/callback signatures |
-| `tools/lsp/common/document_cache.rs` | Document caching and compilation |
-| `tools/lsp/common/rename_component.rs` | Rename of components, structs, enums, properties, callbacks, functions |
-| `tools/lsp/common/host_language_search.rs` | Cross-language rename: walks workspace files to replace matching Rust/C++ accessor identifiers |
+| `internal/editor-preview/common/document_cache.rs` | Document caching and compilation |
+| `internal/editor-preview/common/rename_component.rs` | Rename of components, structs, enums, properties, callbacks, functions |
+| `tools/lsp/host_language_search.rs` | Cross-language rename: walks workspace files to replace matching Rust/C++ accessor identifiers |
 | `internal/compiler/generator/accessor_names.rs` | Shared name mapping for Rust/C++ property/callback/function accessors (used by both codegen and the LSP scanner) |
 | `tools/lsp/preview.rs` | Live preview engine |
 | `tools/lsp/fmt/` | Code formatter |
@@ -309,7 +309,7 @@ pub fn goto_definition(
 
 ## Rename
 
-Rename support lives in `tools/lsp/common/rename_component.rs` and is
+Rename support lives in `internal/editor-preview/common/rename_component.rs` and is
 dispatched from the `textDocument/rename` handler in `language.rs`. It
 handles components, structs, enums, internal/export names, properties,
 callbacks, and functions through a single `DeclarationNode::rename`
@@ -320,8 +320,8 @@ sources.
 
 Renaming a public property, callback, or function can also search and replace
 its generated Rust/C++ accessors in workspace files.
-See `tools/lsp/common/rename_component.rs` for the rename flow and
-`tools/lsp/common/host_language_search.rs` for the workspace search.
+See `internal/editor-preview/common/rename_component.rs` for the rename flow and
+`tools/lsp/host_language_search.rs` for the workspace search.
 
 ## Live Preview
 
@@ -489,7 +489,7 @@ with_lookup_ctx(document_cache, node, Some(offset), |ctx| {
 ### Finding Element at Position
 
 `element_at_position` is a method on `DocumentCache`
-(`tools/lsp/common/document_cache.rs`), not a free function:
+(`internal/editor-preview/common/document_cache.rs`), not a free function:
 
 ```rust
 impl DocumentCache {
