+++ internal/editor-preview/LICENSES/GPL-3.0-only.txt
@@ -1,0 +1,1 @@
+../../../LICENSES/GPL-3.0-only.txt
