+++ internal/editor-preview/LICENSES/LicenseRef-Slint-Royalty-free-2.0.md
@@ -1,0 +1,1 @@
+../../../LICENSES/LicenseRef-Slint-Royalty-free-2.0.md
