+++ internal/editor-preview/LICENSES/LicenseRef-Slint-Software-3.0.md
@@ -1,0 +1,1 @@
+../../../LICENSES/LicenseRef-Slint-Software-3.0.md
