+++ internal/editor-preview/common/text_edit.rs
