+++ internal/editor-preview/lib.rs
@@ -1,0 +1,8 @@
+// Copyright © SixtyFPS GmbH <info@slint.dev>
+// SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-Royalty-free-2.0 OR LicenseRef-Slint-Software-3.0
+
+#![doc = include_str!("README.md")]
+#![deny(clippy::print_stderr, clippy::print_stdout, clippy::disallowed_methods)]
+
+pub mod common;
+pub mod util;
